Refactor MemberController, memberId로 int 대신 long 사용

+++ src/main/java/kr/co/khedu/fitroutine/member/mapper/MemberMapper.java
@@ -6,7 +6,7 @@ import org.apache.ibatis.annotations.Mapper;
 
 @Mapper
 public interface MemberMapper {
-    Member findMemberById(int memberId);
+    Member findMemberById(long memberId);
 
-    MemberProfile getMemberProfile(int memberId);
+    MemberProfile getMemberProfile(long memberId);
 }
